utiles/app/requirements.txt
azureml==0.2.7
Flask==2.0.3
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.0.1
requests==2.27.1
keras==2.8.0